// src/models/MovieDetails.tsx
export interface MovieDetails {
    title: string;
    episode_id: number;
    opening_crawl: string;
    director: string;
    producer: string;
    release_date: string;
    noOfCharacters: number;
    planets: string[];
    noOfStarships: number;
    noOfVehicles: number;
    noOfSpecies: number;
}